* Added ordering setting * Added admin hook

+++ trunk/lnx.milanin.com/egroupware/egroupware/vanilla/inc/class.sovanilla.inc.php
@@ -141,11 +141,13 @@
               }
               function write_disc_watchers($disc_watchers=Array())
               {
-                foreach ( array_keys($disc_watchers) as $disc_id)
-                {
-                  if ($disc_watchers[$disc_id]==0){
-                    $query="DELETE FROM LUM_UserBookmark WHERE UserID=".$this->owner." and DiscussionID=".$disc_id;
-                    $this->db->query($query,__LINE__,__FILE__);
+                if ( is_array($disc_watchers)){
+                  foreach ( array_keys($disc_watchers) as $disc_id)
+                  {
+                    if ($disc_watchers[$disc_id]==0){
+                      $query="DELETE FROM LUM_UserBookmark WHERE UserID=".$this->owner." and DiscussionID=".$disc_id;
+                      $this->db->query($query,__LINE__,__FILE__);
+                    }
                   }
                 }
                 return True;
@@ -154,5 +156,44 @@
               {
                 $query='DELETE FROM LUM_UserCategoryWatch Where UserID ='.$this->owner;
                 $this->db->query($query,__LINE__,__FILE__);
+                $this->db->next_record();
+                return True;
+              }
+              function get_settings(){
+                $query='SELECT Settings FROM LUM_User Where UserID ='.$this->owner;
+                $this->db->query($query,__LINE__,__FILE__);
+                $this->db->next_record();
+                $settings = $this->db->f('Settings');
+                if ( isset( $settings ) )
+                {
+                  return $this->UnserializeAssociativeArray($this->db->f('Settings'));
+                }else{
+                  return Array('HtmlOn' => 1);
+                }
+              }
+              function save_settings($settings)
+              {
+                if (!isset($settings['HtmlOn']) && $settings['HtmlOn'] != 1) $settings['HtmlOn'] = 1;
+//                 echo '<!-- serialized: ['.$this->SerializeArray($settings).'] -->';
+                $query='UPDATE LUM_User SET Settings=\''.$this->SerializeArray($settings).'\' Where UserID='.$this->owner;
+                $this->db->query($query,__LINE__,__FILE__);
+                return True;
+              }
+              function UnserializeAssociativeArray($InSerialArray) {
+                $aReturn = array();
+                if ($InSerialArray != "" && !is_array($InSerialArray)) {
+                  $aReturn = unserialize($InSerialArray);
+                }
+                return $aReturn;	
+              }
+              function SerializeArray($InArray) {
+                $sReturn = "";
+                if (is_array($InArray)) {
+                        if (count($InArray) > 0) {
+                                $sReturn = serialize($InArray);
+                                $sReturn = addslashes($sReturn);
+                        }
+                }
+                return $sReturn;
               }
 }
